requirements.txt
discord.py==2.5.2
openai==1.92.2
pytest==8.4.1
python-dotenv==1.0.0
python-fasthtml==0.12.20
requests==2.32.4